pr feedback: move stuff to default trait
