+++ core/src/repair/standard_repair_handler.rs
@@ -1,17 +1,8 @@
 use {
-    super::{
-        repair_handler::RepairHandler,
-        repair_response,
-        serve_repair::{AncestorHashesResponse, MAX_ANCESTOR_RESPONSES},
-    },
-    bincode::serialize,
+    super::{repair_handler::RepairHandler, repair_response},
     solana_clock::Slot,
-    solana_ledger::{
-        ancestor_iterator::{AncestorIterator, AncestorIteratorWithHash},
-        blockstore::Blockstore,
-        shred::Nonce,
-    },
-    solana_perf::packet::{PacketBatch, PacketBatchRecycler, PinnedPacketBatch},
+    solana_ledger::{blockstore::Blockstore, shred::Nonce},
+    solana_perf::packet::{Packet, PacketBatch, PacketBatchRecycler, PinnedPacketBatch},
     std::{net::SocketAddr, sync::Arc},
 };
 
@@ -26,61 +17,24 @@ impl StandardRepairHandler {
 }
 
 impl RepairHandler for StandardRepairHandler {
-    fn run_window_request(
+    fn blockstore(&self) -> &Blockstore {
+        &self.blockstore
+    }
+
+    fn repair_response_packet(
         &self,
-        recycler: &PacketBatchRecycler,
-        from_addr: &SocketAddr,
         slot: Slot,
         shred_index: u64,
+        dest: &SocketAddr,
         nonce: Nonce,
-    ) -> Option<PacketBatch> {
-        // Try to find the requested index in one of the slots
-        let packet = repair_response::repair_response_packet(
+    ) -> Option<Packet> {
+        repair_response::repair_response_packet(
             self.blockstore.as_ref(),
             slot,
             shred_index,
-            from_addr,
+            dest,
             nonce,
-        )?;
-        Some(
-            PinnedPacketBatch::new_unpinned_with_recycler_data(
-                recycler,
-                "run_window_request",
-                vec![packet],
-            )
-            .into(),
         )
-    }
-
-    fn run_highest_window_request(
-        &self,
-        recycler: &PacketBatchRecycler,
-        from_addr: &SocketAddr,
-        slot: Slot,
-        highest_index: u64,
-        nonce: Nonce,
-    ) -> Option<PacketBatch> {
-        // Try to find the requested index in one of the slots
-        let meta = self.blockstore.meta(slot).ok()??;
-        if meta.received > highest_index {
-            // meta.received must be at least 1 by this point
-            let packet = repair_response::repair_response_packet(
-                self.blockstore.as_ref(),
-                slot,
-                meta.received - 1,
-                from_addr,
-                nonce,
-            )?;
-            return Some(
-                PinnedPacketBatch::new_unpinned_with_recycler_data(
-                    recycler,
-                    "run_highest_window_request",
-                    vec![packet],
-                )
-                .into(),
-            );
-        }
-        None
     }
 
     fn run_orphan(
@@ -110,42 +64,5 @@ impl RepairHandler for StandardRepairHandler {
             res.push(packet);
         }
         (!res.is_empty()).then_some(res.into())
-    }
-
-    fn run_ancestor_hashes(
-        &self,
-        recycler: &PacketBatchRecycler,
-        from_addr: &SocketAddr,
-        slot: Slot,
-        nonce: Nonce,
-    ) -> Option<PacketBatch> {
-        let ancestor_slot_hashes = if self.blockstore.is_duplicate_confirmed(slot) {
-            let ancestor_iterator = AncestorIteratorWithHash::from(
-                AncestorIterator::new_inclusive(slot, self.blockstore.as_ref()),
-            );
-            ancestor_iterator.take(MAX_ANCESTOR_RESPONSES).collect()
-        } else {
-            // If this slot is not duplicate confirmed, return nothing
-            vec![]
-        };
-        let response = AncestorHashesResponse::Hashes(ancestor_slot_hashes);
-        let serialized_response = serialize(&response).ok()?;
-
-        // Could probably directly write response into packet via `serialize_into()`
-        // instead of incurring extra copy in `repair_response_packet_from_bytes`, but
-        // serialize_into doesn't return the written size...
-        let packet = repair_response::repair_response_packet_from_bytes(
-            serialized_response,
-            from_addr,
-            nonce,
-        )?;
-        Some(
-            PinnedPacketBatch::new_unpinned_with_recycler_data(
-                recycler,
-                "run_ancestor_hashes",
-                vec![packet],
-            )
-            .into(),
-        )
     }
 }
